Switch to f32

+++ src/zig/root.zig
@@ -53,10 +53,10 @@ pub fn get_buf(arr: *f64JArray) []f64 {
 
 pub export fn populateNoiseArray(
     noiseArray: [*]f64,
-    xOffset: f64, yOffset: f64, zOffset: f64,
+    xOffset: f32, yOffset: f32, zOffset: f32,
     xSize: i32, ySize: i32, zSize: i32,
-    xScale: f64, yScale: f64, zScale: f64,
-    noiseScale: f64, seed: i64) void {
+    xScale: f32, yScale: f32, zScale: f32,
+    noiseScale: f32, seed: i64) void {
     @setFloatMode(.optimized);
     const bLen = @as(usize, @intCast(xSize * ySize * zSize));
     var buffer = noiseArray[0..bLen];
@@ -66,16 +66,16 @@ pub export fn populateNoiseArray(
     const zMax = @as(usize, @intCast(zSize));
 
     for (0..xMax) |px| {
-        const fx = @as(f64, @floatFromInt(px)) * xScale + xOffset;
+        const fx = @as(f32, @floatFromInt(px)) * xScale + xOffset;
 
         for (0..yMax) |py| {
-            const fy = @as(f64, @floatFromInt(py)) * yScale + yOffset;
+            const fy = @as(f32, @floatFromInt(py)) * yScale + yOffset;
 
             for (0..zMax) |pz| {
-                const fz = @as(f64, @floatFromInt(pz)) * zScale + zOffset;
+                const fz = @as(f32, @floatFromInt(pz)) * zScale + zOffset;
                 const bidx = py + px * yMax + pz * xMax * yMax;
 
-                buffer[bidx] = opensimplex.noise3_ImproveXZ(seed, fx, fy, fz) * noiseScale;
+                buffer[bidx] = opensimplex.noise3_ImproveXZ(seed,fx, fy,fz) * noiseScale;
             }
         }
     }
@@ -83,10 +83,10 @@ pub export fn populateNoiseArray(
 
 pub export fn FNL_populateNoiseArray(
     noiseArray: [*]f64,
-    xOffset: f64, yOffset: f64, zOffset: f64,
+    xOffset: f32, yOffset: f32, zOffset: f32,
     xSize: i32, ySize: i32, zSize: i32,
-    xScale: f64, yScale: f64, zScale: f64,
-    noiseScale: f64, seed: i64) void {
+    xScale: f32, yScale: f32, zScale: f32,
+    noiseScale: f32, seed: i64) void {
     @setFloatMode(.optimized);
     const bLen = @as(usize, @intCast(xSize * ySize * zSize));
     var buffer = noiseArray[0..bLen];
@@ -95,7 +95,7 @@ pub export fn FNL_populateNoiseArray(
     const yMax = @as(usize, @intCast(ySize));
     const zMax = @as(usize, @intCast(zSize));
 
-    const NoiseGen = fastnoiselite.Noise(f64);
+    const NoiseGen = fastnoiselite.Noise(f32);
     const generator: NoiseGen = .{
         .seed = @truncate(seed),
         .frequency = 1,
@@ -104,16 +104,16 @@ pub export fn FNL_populateNoiseArray(
     };
 
     for (0..xMax) |px| {
-        const fx = @as(f64, @floatFromInt(px)) * xScale + xOffset;
+        const fx = @as(f32, @floatFromInt(px)) * xScale + xOffset;
 
         for (0..yMax) |py| {
-            const fy = @as(f64, @floatFromInt(py)) * yScale + yOffset;
+            const fy = @as(f32, @floatFromInt(py)) * yScale + yOffset;
 
             for (0..zMax) |pz| {
-                const fz = @as(f64, @floatFromInt(pz)) * zScale + zOffset;
+                const fz = @as(f32, @floatFromInt(pz)) * zScale + zOffset;
                 const bidx = py + px * yMax + pz * xMax * yMax;
 
-                buffer[bidx] = generator.genNoise3D(fx, fy, fz) * noiseScale;
+                buffer[bidx] = generator.genNoise3D(fx, fy,fz) * noiseScale;
             }
         }
     }
